Conversion to angular app
