+++ src/app/app.component.ts
@@ -1,14 +1,14 @@
-import { Component } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
+import {Component} from '@angular/core';
+import {RouterOutlet} from '@angular/router';
 import {BannerComponent} from "./components/banner/banner.component";
-import {CueInputComponent} from "./components/cue-input/cue-input.component";
 
 @Component({
-  selector: 'app-root',
-  imports: [RouterOutlet, BannerComponent, CueInputComponent],
-  templateUrl: './app.component.html',
-  styleUrl: './app.component.scss',
+    selector: 'app-root',
+    imports: [RouterOutlet, BannerComponent],
+    templateUrl: './app.component.html',
+    styleUrl: './app.component.scss',
+    standalone: true
 })
 export class AppComponent {
-  title = 'cueter';
+    title = 'cueter';
 }
