+++ src/app/app.routes.ts
@@ -1,3 +1,9 @@
 import { Routes } from '@angular/router';
+import {ConverterComponent} from "./pages/converter/converter.component";
 
-export const routes: Routes = [];
+export const routes: Routes = [
+    {
+        path: '',
+        component: ConverterComponent
+    }
+];
