+++ src/app/components/banner/banner.component.html
@@ -1,3 +1,3 @@
-<header class="container md:px-4 md:py-4 px-2 py-2">
+<header class="md:px-4 md:py-4 px-2 py-2 bg-emerald-600">
     <h1>Cueter</h1>
 </header>
