--- src/app/components/cue-input/cue-input.component.html
@@ -1,6 +1,0 @@
-<form [formGroup]="cueGroup" (ngSubmit)="saveForm()">
-    <label for="cue-input">Cue text</label>
-<textarea id="cue-input" type="text" formControlName="cueText">
-</textarea>
-    <input type="submit" name="Toto" />
-</form>
