+++ src/app/pages/converter/converter.component.html
@@ -1,0 +1,19 @@
+<main class="md:container md:mx-auto">
+    <div class="flex flex-row">
+        <div class="basis-1/2">
+            <form [formGroup]="cueForm" (ngSubmit)="saveForm()" class="flex flex-col p-2">
+                <label for="cue-input" class="bg-emerald-900 text-slate-200">Cue text</label>
+                <textarea id="cue-input" type="text" formControlName="input" class="border-solid border-2">
+</textarea>
+                <label for="cue-track-template" class="bg-emerald-900 text-slate-200">Template</label>
+                <input id="cue-track-template" type="text" formControlName="template"
+                       class="border-solid border-2"/>
+                <input type="submit" value="convert" class="mt-3 p-2 bg-emerald-900 text-slate-200 rounded-md"/>
+            </form>
+        </div>
+        <div class="basis-1/2">
+            <h2>Result</h2>
+            <textarea class="bg-slate-50" readonly>{{output()}}</textarea>
+        </div>
+    </div>
+</main>
